+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - [Optimize/optimize annotation.](https://github.com/Tencent/spring-cloud-tencent/pull/672)
 - [Optimize: Register the service with the ProviderAPI#registerInstance method.](https://github.com/Tencent/spring-cloud-tencent/pull/686)
 - [docs:update PR template.](https://github.com/Tencent/spring-cloud-tencent/pull/690)
+- [enhancement: revert default value when the field is deleted in @ConfigurationProperties bean](https://github.com/Tencent/spring-cloud-tencent/issues/681)